cleanup redundancy between DEFAULT_SLIPPAGE
